Split out custom image viewer into separate file
